Implement new stream id allocator and grim reaper for slave corpses
Implement simple bitmap based id allocator which allocates the lowest
available id.  Reaping thread is added so that stream id is freed only
after the slave is completely dead.  This is to prepare for new mmap
implementation.

Note that as id 0 is reserved, this reduces max number of streams by
one.